src/requirements.txt
torch
torchvision
opencv-python
rasterio
lmdb
tqdm
opencv-torchvision-transforms-yuzhiyang
matplotlib
scikit-learn